+++ app/src/main/java/it/unibo/unibodget/model/dashboard/impl/SafeBudgetStrategy.java
@@ -2,7 +2,7 @@ package it.unibo.unibodget.model.dashboard.impl;
 
 import it.unibo.unibodget.model.dashboard.api.BudgetAlertStrategy;
 import it.unibo.unibodget.model.dashboard.api.BudgetStatus;
-
+import java.math.BigDecimal;
 /**
  * Strategy representing a safe budget condition.
  * This strategy is selected when the current value is clearly below the warning threshold.
@@ -13,7 +13,7 @@ public final class SafeBudgetStrategy implements BudgetAlertStrategy {
      * {@inheritDoc}
      */
     @Override
-    public BudgetStatus evaluate(final double currentValue, final double limitValue) {
+    public BudgetStatus evaluate(final BigDecimal currentValue, final BigDecimal limitValue) {
         return BudgetStatus.SAFE;
     }
 }
