+++ app/src/main/java/it/unibo/unibodget/model/dashboard/impl/WarningBudgetStrategy.java
@@ -1,8 +1,9 @@
 package it.unibo.unibodget.model.dashboard.impl;
 
+import java.math.BigDecimal;
+
 import it.unibo.unibodget.model.dashboard.api.BudgetAlertStrategy;
 import it.unibo.unibodget.model.dashboard.api.BudgetStatus;
-
 /**
  * Strategy representing an approaching-limit budget condition.
  */
@@ -12,7 +13,8 @@ public final class WarningBudgetStrategy implements BudgetAlertStrategy {
      * {@inheritDoc}
      */
     @Override
-    public BudgetStatus evaluate(final double currentValue, final double limitValue) {
+    public BudgetStatus evaluate(final BigDecimal currentValue, final BigDecimal
+         limitValue) {
         return BudgetStatus.WARNING;
     }
 }
